requirements.txt
SQLAlchemy~=2.0.36
fastapi~=0.115.5
starlette~=0.41.3
PyJWT~=2.10.1
uvicorn~=0.32.1
pydantic~=2.10.2
psycopg2~=2.9.10
python-multipart~=0.0.20